+++ src/actions/start.action.ts
@@ -83,6 +83,7 @@ export async function startAction() {
   answers.todoConfig.forEach((c: string) => {
     todoConfigMap[c] = true;
   });
+  todoConfigMap[answers.gitLabProject.id] = true;
   const template = readFileSync(untildify(CONFIG.ToDoTemplate), {
     encoding: "utf-8",
   });