+++ src/classes/daily-progress.class.ts
@@ -1,6 +1,6 @@
-import untildify from 'untildify';
-import { BaseFileRef } from './base-file-ref.class';
-import { CONFIG } from '../config';
+import untildify from "untildify";
+import { CONFIG } from "../config";
+import { BaseFileRef } from "./base-file-ref.class";
 
 export class DailyProgress extends BaseFileRef {
   protected get path() {
@@ -10,7 +10,7 @@ export class DailyProgress extends BaseFileRef {
   public addProgressToBuffer(dailyProgressString: string) {
     const content = this.readFile();
     const updatedDpContent = content.replace(
-      '## Buffer',
+      "## Buffer",
       `## Buffer\n    ${dailyProgressString}`
     );
     this.writeFile(updatedDpContent);
@@ -18,17 +18,17 @@ export class DailyProgress extends BaseFileRef {
 
   public getRecordByDay(day: string) {
     const content = this.readFile();
-    const matchResult = content.match(new RegExp(`(### ${day}.*?)\n###`, 's'));
+    const matchResult = content.match(new RegExp(`(### ${day}.*?)\n###`, "s"));
     if (matchResult) {
       const record = matchResult[1];
       if (/2\. Today\n3\./.test(record)) {
-        console.log('Today content is empty.');
+        console.log("Today content is empty.");
         return null;
       } else {
         return record;
       }
     } else {
-      console.log('DP record does not exist.');
+      console.log("DP record does not exist.");
       return null;
     }
   }